+++ test/libsolidity/semanticTests/dirty_calldata_dynamic_array.sol
@@ -1,6 +1,6 @@
 contract C {
     function f(int16[] calldata a) external returns (bool correct) {
-        uint32 x = uint32(a[1]);
+        uint32 x = uint32(uint16(a[1]));
         uint r;
         assembly {
             r := x
