+++ test/libsolidity/semanticTests/types/type_conversion_cleanup.sol
@@ -1,8 +1,8 @@
 contract Test {
-    function test() public returns (uint ret) { return uint(address(uint128(type(uint200).max))); }
+    function test() public returns (uint ret) { return uint(uint160(address(uint160(type(uint200).max)))); }
 }
 // ====
 // compileViaYul: also
 // compileToEwasm: also
 // ----
-// test() -> 0xffffffffffffffffffffffffffffffff
+// test() -> 0xffffffffffffffffffffffffffffffffffffffff
